app/app/requirements.txt
Pillow==9.0.1
tensorflow==2.7.1